+++ projects/evo-ui-kit/src/lib/common/types/evo-size-keys.ts
@@ -1,3 +1,0 @@
-import { EvoSize } from '../enums/index';
-
-export type EvoSizeKeys = keyof typeof EvoSize;