+++ projects/evo-ui-kit/src/lib/components/evo-badge/evo-badge.module.ts
@@ -1,9 +1,7 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { EvoBadgeComponent } from './evo-badge.component';
-import { EvoUiKitModule } from '@evo/ui-kit';
-
-
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {EvoBadgeComponent} from './evo-badge.component';
+import {EvoUiKitModule} from '../../evo-ui-kit.module';
 
 @NgModule({
   declarations: [EvoBadgeComponent],