+++ projects/evo-ui-kit/src/lib/components/evo-badge/types/index.ts
@@ -1,0 +1,2 @@
+export * from './evo-badge-color';
+export * from './evo-badge-size';