+++ src/stories/components/evo-badge.stories.ts
@@ -1,6 +1,5 @@
 import { moduleMetadata, storiesOf } from '@storybook/angular';
 import { EvoBadgeModule } from '../../../projects/evo-ui-kit/src/lib/components/evo-badge';
-import { EvoColor, EvoSize } from '../../../projects/evo-ui-kit/src/lib/common/enums';
 
 storiesOf('Components/Badge', module)
     .addDecorator(
@@ -32,21 +31,23 @@ storiesOf('Components/Badge', module)
         props: {
             colorsList: [
                 'DEFAULT_NO_COLOR',
-                EvoColor.success,
-                EvoColor.error,
-                EvoColor.iconDark,
-                EvoColor.graph1,
-                EvoColor.graph2,
-                EvoColor.graph3,
-                EvoColor.graph4,
-                EvoColor.graph5,
-                EvoColor.graph6,
-                EvoColor.graph7,
-                EvoColor.graph8,
-                EvoColor.graph9,
-                EvoColor.graph10,
-            ]
-        }
+                'success',
+                'error',
+                'icon-dark',
+                'graph-1',
+                'graph-2',
+                'graph-3',
+                'graph-4',
+                'graph-5',
+                'graph-6',
+                'graph-7',
+                'graph-8',
+                'graph-9',
+                'graph-10',
+                'grey',
+                'primary',
+            ],
+        },
     }))
     .add('with size', () => ({
         styleUrls: ['../../assets/scss/story-global.scss'],
@@ -59,16 +60,53 @@ storiesOf('Components/Badge', module)
         <div class="story-container">
             <h2>Badges and available sizes</h2>
             <div class="story-section">
-                    <div class="badge-container" *ngFor="let size of sizesList">
-                        <evo-badge [size]="size">{{size}}</evo-badge>
-                    </div>
+                <div class="badge-container" *ngFor="let size of sizesList">
+                    <evo-badge [size]="size">{{size}}</evo-badge>
+                </div>
             </div>
         </div>
         `,
         props: {
             sizesList: [
-               'DEFAULT',
-               EvoSize.small,
-            ]
-        }
+               'normal',
+               'small',
+            ],
+        },
+    }))
+    .add('with fixed width', () => ({
+        styleUrls: ['../../assets/scss/story-global.scss'],
+        template: `
+        <style>
+            .badge-container {
+                width: 100%;
+                margin-bottom: 10px;
+            }
+        </style>
+        <div class="story-container">
+            <h2>Badges and custom widths</h2>
+            <div class="story-section">
+                <div class="badge-container">
+                    <evo-badge [width.px]="50">50px</evo-badge>
+                </div>
+                <div class="badge-container">
+                    <evo-badge [width.%]="20">20%</evo-badge>
+                </div>
+            </div>
+        </div>
+        `,
+    }))
+    .add('with multiline', () => ({
+        styleUrls: ['../../assets/scss/story-global.scss'],
+        template: `
+        <div class="story-container">
+            <h2>Badge multiline property</h2>
+            <div class="story-section">
+                <evo-badge multiline="true" [width.px]="85">
+                    I <3️ Angular
+                    I <3️ Angular
+                    I <3️ Angular
+                </evo-badge>
+            </div>
+        </div>
+        `,
     }));